requirements.txt
discord.py
selenium
requests
redis
pyyaml
pymongo
pydantic-settings